+++ qord/__init__.py
@@ -27,6 +27,7 @@ from qord.models.guilds import *
 from qord.models.guild_members import *
 from qord.models.roles import *
 from qord.models.messages import *
+from qord.models.emojis import *
 from qord.dataclasses.allowed_mentions import *
 from qord.dataclasses.embeds import *
 from qord.dataclasses.files import *